chore(databases): move all load data related components to their own file; adjust naming to be move specific
