+++ packages/databases-collections/src/components/databases.tsx
@@ -3,16 +3,14 @@ import { connect } from 'react-redux';
 import {
   Banner,
   BannerVariant,
-  Button,
   EmptyContent,
   Link,
   css,
   spacing,
-  Icon,
 } from '@mongodb-js/compass-components';
 import { DatabasesList } from '@mongodb-js/databases-collections-list';
 import { usePreference } from 'compass-preferences-model/provider';
-import { AddDataZeroGraphic, ZeroGraphic } from './zero-graphic';
+import { ZeroGraphic } from './zero-graphic';
 import type { Database, DatabasesState } from '../modules/databases';
 import {
   createDatabase,
@@ -26,6 +24,10 @@ import {
   type TrackFunction,
 } from '@mongodb-js/compass-telemetry/provider';
 import toNS from 'mongodb-ns';
+import {
+  LoadSampleDataZeroBanner,
+  LoadSampleDataZeroState,
+} from './load-sample-data';
 
 const errorContainerStyles = css({
   padding: spacing[400],
@@ -61,125 +63,6 @@ function NonGenuineZeroState() {
         }
       />
     </div>
-  );
-}
-
-function buildAddDataUrl(projectId: string, clusterName: string) {
-  const url = new URL(
-    `/v2/${projectId}#/addData/${encodeURIComponent(clusterName)}/load`,
-    window.location.origin
-  );
-  return url.toString();
-}
-
-const addDataContainerStyles = css({
-  width: '100%',
-  padding: spacing[400],
-});
-
-const addDataActionsStyles = css({
-  display: 'flex',
-  gap: spacing[300],
-});
-
-function AddDataZeroState({
-  projectId,
-  clusterName,
-  canCreateDatabase,
-  onCreateDatabase,
-}: {
-  projectId: string;
-  clusterName: string;
-  canCreateDatabase: boolean;
-  onCreateDatabase: () => void;
-}) {
-  return (
-    <div className={addDataContainerStyles} data-testid="add-data-zero-state">
-      <EmptyContent
-        icon={AddDataZeroGraphic}
-        title="Looks like your cluster is empty"
-        subTitle={
-          canCreateDatabase ? (
-            <>
-              Create database or load sample data to your cluster to quickly get
-              started experimenting with data in MongoDB.
-            </>
-          ) : (
-            <>
-              You can load sample data to quickly get started experimenting with
-              data in MongoDB.
-            </>
-          )
-        }
-        callToActionLink={
-          <div className={addDataActionsStyles}>
-            {canCreateDatabase && (
-              <Button variant="default" onClick={onCreateDatabase}>
-                Create database
-              </Button>
-            )}
-            <Button
-              variant="primary"
-              href={buildAddDataUrl(projectId, clusterName)}
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Load sample dataset
-            </Button>
-          </div>
-        }
-      />
-    </div>
-  );
-}
-
-const addDataBannerContent = css({
-  display: 'flex',
-  alignItems: 'center',
-});
-
-const addDataBannerButtonStyles = css({
-  marginLeft: 'auto',
-  whiteSpace: 'nowrap',
-  textOverflow: 'ellipsis',
-});
-
-const EmptyElement = () => null;
-
-function AddDataZeroBanner({
-  projectId,
-  clusterName,
-}: {
-  projectId: string;
-  clusterName: string;
-}) {
-  return (
-    <Banner image={<EmptyElement></EmptyElement>}>
-      <div className={addDataBannerContent}>
-        <span>
-          Working with MongoDB is easy, but first you’ll need some data to get
-          started. Sample data is available for loading.
-        </span>
-        <Button
-          className={addDataBannerButtonStyles}
-          onClick={() => {
-            // Leafygreen overrides anchor tag styles inside the banner in a way
-            // that completely breaks the button visuals and there is no good
-            // way for us to hack around it, so instead of a link, we're using a
-            // button and open a url with browser APIs
-            window.open(
-              buildAddDataUrl(projectId, clusterName),
-              '_blank',
-              'noopener noreferrer'
-            );
-          }}
-          leftGlyph={<Icon glyph="Upload"></Icon>}
-          size="small"
-        >
-          Load sample data
-        </Button>
-      </div>
-    </Banner>
   );
 }
 
@@ -232,7 +115,7 @@ const Databases: React.FunctionComponent<DatabasesProps> = ({
     track('Screen', { name: 'databases' });
   }, []);
 
-  const renderBanner = useCallback(() => {
+  const renderLoadSampleDataBanner = useCallback(() => {
     if (
       !atlasMetadata ||
       databases.some((db) => {
@@ -243,10 +126,10 @@ const Databases: React.FunctionComponent<DatabasesProps> = ({
     }
 
     return (
-      <AddDataZeroBanner
+      <LoadSampleDataZeroBanner
         projectId={atlasMetadata.projectId}
         clusterName={atlasMetadata.clusterName}
-      ></AddDataZeroBanner>
+      ></LoadSampleDataZeroBanner>
     );
   }, [databases, atlasMetadata]);
 
@@ -274,12 +157,12 @@ const Databases: React.FunctionComponent<DatabasesProps> = ({
 
   if (atlasMetadata && databases.length === 0) {
     return (
-      <AddDataZeroState
+      <LoadSampleDataZeroState
         projectId={atlasMetadata.projectId}
         clusterName={atlasMetadata.clusterName}
         canCreateDatabase={editable}
         onCreateDatabase={onCreateDatabaseClick}
-      ></AddDataZeroState>
+      ></LoadSampleDataZeroState>
     );
   }
 
@@ -299,7 +182,7 @@ const Databases: React.FunctionComponent<DatabasesProps> = ({
   return (
     <DatabasesList
       databases={databases}
-      renderBanner={renderBanner}
+      renderLoadSampleDataBanner={renderLoadSampleDataBanner}
       {...actions}
     />
   );
